# shared/api-contracts/openapi.yaml
...
openapi: "3.1.0"
info:
  title: "TTBS Customer API"
  description: "Customer-facing API contract for the train ticket booking system."
  version: "1.0"
servers:
- url: "/api"
  description: "Versioned customer API root"
security:
- bearerAuth: []
...
paths:
...
  /api/{version}/stations:
    get:
      tags:
      - "Stations"
      summary: "List stations"
      operationId: "getStations"
      parameters:
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetStationsRequest"
      responses:
        "200":
          description: "Paged station catalog"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PageResponseStationResponse"
        "400":
          description: "Invalid pagination parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/stations/{id}:
    get:
      tags:
      - "Stations"
      summary: "Get a station by id"
      operationId: "getStation"
      parameters:
      - name: "id"
        in: "path"
        description: "Station identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetStationByIdRequest"
      responses:
        "200":
          description: "Station detail"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/StationResponse"
        "404":
          description: "Station not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
  /api/{version}/stations/search:
    get:
      tags:
      - "Stations"
      summary: "Search stations by keyword"
      operationId: "searchStations"
      parameters:
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/SearchStationsRequest"
      responses:
        "200":
          description: "Matching station suggestions"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                type: "array"
                items:
                  $ref: "#/components/schemas/StationSearchResponse"
        "400":
          description: "Invalid station search parameters"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
...
  /api/{version}/payments/{paymentId}:
    get:
      tags:
      - "Payments"
      summary: "Get a payment by id"
      operationId: "getPayment"
      parameters:
      - name: "paymentId"
        in: "path"
        description: "Payment identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetPaymentByIdRequest"
      responses:
        "200":
          description: "Payment detail"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "403":
          description: "Customer cannot access this payment"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Payment not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
...
  /api/{version}/bookings/{bookingId}/payment:
    get:
      tags:
      - "Payments"
      summary: "Get the payment linked to a booking"
      operationId: "getBookingPayment"
      parameters:
      - name: "bookingId"
        in: "path"
        description: "Booking identifier"
        required: true
        schema:
          type: "string"
          format: "uuid"
      - name: "request"
        in: "query"
        required: true
        schema:
          $ref: "#/components/schemas/GetPaymentByBookingIdRequest"
      responses:
        "200":
          description: "Payment linked to the booking"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendResponseObject"
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentResponse"
        "403":
          description: "Customer cannot access this booking payment"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "404":
          description: "Payment or booking not found"
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JsendFailResponse"
        "500":
          description: "Unexpected technical failure."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JsendErrorResponse"
      security:
      - bearerAuth: []
components:
  schemas:
    FailData:
      description: "Shared payload used inside JSend fail responses."
      properties:
        message:
          type: "string"
          description: "Human-readable summary of the failure."
        code:
          $ref: "#/components/schemas/ErrorCode"
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/Violation"
    ErrorCode:
      type: "string"
      description: "Machine-readable error code carried by shared fail responses."
      enum:
      - "VALIDATION_ERROR"
      - "CURSOR_INVALID"
      - "ACCESS_DENIED"
      - "USER_EMAIL_ALREADY_EXISTS"
      - "USER_INVALID_CREDENTIALS"
      - "USER_INVALID_REFRESH_TOKEN"
      - "USER_NOT_FOUND"
      - "USER_CANNOT_BULK_DELETE_SELF"
      - "USER_HAS_ACTIVE_BOOKINGS"
      - "BOOKING_NOT_FOUND"
      - "BOOKING_CANNOT_CONFIRM"
      - "BOOKING_ALREADY_CANCELLED"
      - "TRAIN_NUMBER_ALREADY_EXISTS"
      - "TRAIN_NOT_FOUND"
      - "TRAIN_IN_USE"
      - "SEAT_NOT_FOUND"
      - "SEAT_NUMBER_ALREADY_EXISTS"
      - "SEAT_NOT_AVAILABLE"
      - "SEAT_IN_USE"
      - "SCHEDULED_TRIP_NOT_FOUND"
      - "COACH_NOT_FOUND"
      - "COACH_CAR_NUMBER_ALREADY_EXISTS"
      - "COACH_TRAIN_NOT_FOUND"
      - "COACH_IN_USE"
      - "COACH_CAR_NUMBERS_ALREADY_EXIST"
      - "COACH_DUPLICATE_CAR_NUMBERS_IN_REQUEST"
      - "SEAT_NUMBERS_ALREADY_EXIST"
      - "SEAT_DUPLICATE_SEAT_NUMBERS_IN_REQUEST"
      - "ROUTE_TEMPLATE_NOT_FOUND"
      - "STATION_NOT_FOUND"
      - "STATION_CODE_ALREADY_EXISTS"
      - "STATION_IN_USE"
      - "PAYMENT_NOT_FOUND"
      - "PAYMENT_ALREADY_PROCESSED"
      - "PAYMENT_REFUND_FAILED"
    Violation:
      description: "Field-level validation issue inside a fail response."
      properties:
        field:
          type: "string"
          description: "Request field that failed validation."
        message:
          type: "string"
          description: "Human-readable validation message."
        code:
          $ref: "#/components/schemas/ViolationCode"
    ViolationCode:
      type: "string"
      description: "Machine-readable validation category attached to a field violation."
      enum:
      - "REQUIRED"
      - "INVALID_FORMAT"
      - "TOO_SHORT"
      - "TOO_LONG"
      - "INVALID_VALUE"
    JsendFailResponse:
      type: "object"
      description: "Shared JSend fail envelope used for validation and domain errors."
      properties:
        status:
          type: "string"
          enum:
          - "fail"
        data:
          $ref: "#/components/schemas/FailData"
    JsendErrorResponse:
      type: "object"
      description: "Shared JSend error envelope used for unexpected technical failures."
      properties:
        status:
          type: "string"
          enum:
          - "error"
        message:
          type: "string"
          description: "Human-readable technical failure message."
...
    JsendResponseObject:
      type: "object"
      description: "Runtime JSend response envelope used by the API implementation."
      properties:
        status:
          type: "string"
          description: "JSend status value."
          enum:
          - "success"
          - "fail"
          - "error"
        data:
          description: "Payload returned for the response status."
        message:
          type: "string"
          description: "Human-readable technical error message for error responses."
...
    GetStationsRequest:
      type: "object"
      description: "Offset-based pagination query for stations."
      properties:
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Number of stations per page."
          example: 20
          maximum: 100
          minimum: 1
    GetStationByIdRequest:
      description: "No additional query parameters are accepted when fetching a station."
    SearchStationsRequest:
      type: "object"
      description: "Station auto-complete query."
      properties:
        q:
          type: "string"
          description: "Free-text station search term."
          example: "ha noi"
        limit:
          type: "integer"
          format: "int32"
          description: "Maximum number of station suggestions to return."
          example: 10
          maximum: 20
          minimum: 1
...
    GetPaymentByIdRequest:
      description: "No additional query parameters are accepted when fetching a payment\
        \ by id."
...
    GetPaymentByBookingIdRequest:
      description: "No additional query parameters are accepted when fetching the\
        \ payment for a booking."
...
    PageResponseStationResponse:
      description: "Offset-based pagination wrapper."
      properties:
        content:
          type: "array"
          items:
            $ref: "#/components/schemas/StationResponse"
        page:
          type: "integer"
          format: "int32"
          description: "Zero-based page index."
          example: 0
          minimum: 0
        size:
          type: "integer"
          format: "int32"
          description: "Requested page size."
          example: 20
          minimum: 1
        hasNext:
          type: "boolean"
          description: "Whether a subsequent page exists."
        hasPrevious:
          type: "boolean"
          description: "Whether a previous page exists."
        total:
          type: "integer"
          format: "int64"
          description: "Total number of matching items across all pages."
          example: 125
          minimum: 0
    StationResponse:
      description: "Station summary."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Station identifier."
        code:
          type: "string"
          description: "Station code."
          example: "HNO"
        name:
          type: "string"
          description: "Station name."
          example: "Ha Noi"
        city:
          type: "string"
          description: "Station city."
          example: "Ha Noi"
        createdAt:
          type: "string"
          format: "date-time"
          description: "Station creation timestamp."
    StationSearchResponse:
      description: "Station suggestion returned by station search."
      properties:
        id:
          type: "string"
          format: "uuid"
          description: "Station identifier."
        code:
          type: "string"
          description: "Station code."
          example: "HNO"
        name:
          type: "string"
          description: "Station name."
          example: "Ha Noi"
        city:
          type: "string"
          description: "Station city."
          example: "Ha Noi"
...
    PaymentResponse:
      description: "Payment resource exposed to customers."
      properties:
        paymentId:
          type: "string"
          format: "uuid"
          description: "Payment identifier."
        bookingId:
          type: "string"
          format: "uuid"
          description: "Booking identifier linked to the payment."
        status:
          type: "string"
          description: "Payment lifecycle status."
          enum:
          - "PENDING"
          - "PAID"
          - "CANCELLED"
          - "FAILED"
          - "REFUNDED"
        checkoutUrl:
          type: "string"
          format: "uri"
          description: "Hosted checkout URL when customer action is still required."
        amount:
          type: "number"
          description: "Payment amount in major currency units."
          example: 650000
        currency:
          type: "string"
          description: "ISO-like currency code."
          example: "VND"
...
  securitySchemes:
    bearerAuth:
      type: "http"
      scheme: "bearer"
      bearerFormat: "JWT"